requirements.txt
pandas
numpy
matplotlib
yfinance
feedparser
jinja2
nltk
python-dateutil
